+++ src/js/base/api.js
@@ -190,7 +190,7 @@ ripe.Ripe.prototype.signinPid = function(token, options, callback) {
 
 ripe.Ripe.prototype.signinPidP = function(token, options) {
     return new Promise((resolve, reject) => {
-        this.signinAdmin(token, options, (result, isValid, request) => {
+        this.signinPid(token, options, (result, isValid, request) => {
             isValid ? resolve(result) : reject(new ripe.RemoteError(request, null, result));
         });
     });